feat: restructure, remove uhtml dep
connectedCallback() is now being called in LOL, not in LOLElement
to make it easier for vanilla users to take full control
