+++ package.json
@@ -26,13 +26,11 @@
   "author": "Arturo Castillo Delgado <ac@iconstorm.com>",
   "repository": "iconstorm/lol-element",
   "license": "MIT",
-  "dependencies": {
-    "uhtml": "^2.7.5"
-  },
   "peerDependencies": {
     "htm": "^3",
     "lit-html": ">= 1 <3",
-    "preact": "^10"
+    "preact": "^10",
+    "uhtml": "^3"
   },
   "devDependencies": {
     "@babel/eslint-parser": "^7.13.10",
