+++ src/lit-html/lol.js
@@ -8,6 +8,15 @@ export { html, svg } from 'lit-html'
 
 export class LOL extends LOLElement {
   /**
+   * First render.
+   * Order is important: adoptStyles() must be called AFTER render()
+   */
+  connectedCallback () {
+    this.render()
+    this.adoptStyles()
+  }
+
+  /**
    * Update the DOM.
    */
   render () {
